@@ -207,9 +207,7 @@ func (b *httpBatch) Send() (err error) {
 
 	options.settings["query"] = b.query
 	headers["Content-Type"] = "application/octet-stream"
-	for k, v := range b.conn.headers {
-		headers[k] = v
-	}
+
 	res, err := b.conn.sendStreamQuery(b.ctx, r, &options, headers)
 
 	if res != nil {
