@@ -18,10 +18,11 @@
 package clickhouse
 
 import (
-	"github.com/ClickHouse/clickhouse-go/v2/lib/driver"
-	"github.com/pkg/errors"
+	"errors"
 	"regexp"
 	"time"
+
+	"github.com/ClickHouse/clickhouse-go/v2/lib/driver"
 )
 
 var (
