@@ -20,10 +20,11 @@ package std
 import (
 	"context"
 	"fmt"
-	clickhouse_tests "github.com/ClickHouse/clickhouse-go/v2/tests"
-	"github.com/stretchr/testify/require"
 	"strconv"
 	"testing"
+
+	clickhouse_tests "github.com/ClickHouse/clickhouse-go/v2/tests"
+	"github.com/stretchr/testify/require"
 
 	"github.com/ClickHouse/clickhouse-go/v2"
 	"github.com/stretchr/testify/assert"
@@ -38,15 +39,15 @@ func TestStdTemporaryTable(t *testing.T) {
 			ctx := context.Background()
 			if name == "Http" {
 				ctx = clickhouse.Context(ctx, clickhouse.WithSettings(clickhouse.Settings{
-					"session_id": "test_session",
+					"session_id": "temp_table_test_session",
 				}))
 			}
+
 			conn, err := GetStdDSNConnection(protocol, useSSL, nil)
 			require.NoError(t, err)
-			conn.Exec("DROP TABLE IF EXISTS test_temporary_table")
-			defer func() {
-				conn.Exec("DROP TABLE test_temporary_table")
-			}()
+
+			_, err = conn.Exec("DROP TABLE IF EXISTS test_temporary_table")
+			require.NoError(t, err)
 			const ddl = `CREATE TEMPORARY TABLE test_temporary_table (
 							ID UInt64
 						);`
@@ -78,9 +79,12 @@ func TestStdTemporaryTable(t *testing.T) {
 				require.True(t, ok)
 				assert.Equal(t, int32(60), exception.Code)
 			}
-			require.Equal(t, int(10), count)
+			require.Equal(t, 10, count)
 			require.NoError(t, tx.Commit())
-			assert.NoError(t, conn.Close())
+
+			_, err = conn.Exec("DROP TABLE IF EXISTS test_temporary_table")
+			require.NoError(t, err)
+			require.NoError(t, conn.Close())
 		})
 	}
 }
